Tema13 - Despliegues y funcionalidades avanzadas/Clase39_16-01-2026/api-membresia-drf/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
dj-database-url
requests